+++ R/colors.R
@@ -1,8 +1,3 @@
-# wrapper to get the number of identified clusters:
-count_umap_clusters <- function(seurat_obj) {
-    length(levels(seurat_obj@meta.data[["seurat_clusters"]]))
-}
-
 # get the ggplot colors - important function for 'apotc'
 gg_color_hue <- function(n) {
     hues <- seq(15, 375, length = n + 1)
@@ -55,10 +50,6 @@ extract_and_add_colors <- function(apotc_obj, plot_df) {
   plot_df %>% dplyr::mutate("color" = color_vec)
 }
 
-# in the future should probably make a fake ggplot cluster legend on the right
-# side by inserting scatterplt points? (like in the seurat UMPA plot) But also,
-# the problem is that it becomes inconsistent with the clone size legend :/
-
 #' @noRd
 #' @source https://stackoverflow.com/questions/649454/what-is-the-best-way-to-average-two-colors-that-define-a-linear-gradient#:~:text=NewColor%20%3D%20sqrt((R1%5E2%2BR2%5E2)/2)%2Csqrt((G1%5E2%2BG2%5E2)/2)%2Csqrt((B1%5E2%2BB2%5E2)/2)
 get_average_hex <- function(hex1, hex2) {
@@ -66,4 +57,17 @@ get_average_hex <- function(hex1, hex2) {
     t((grDevices::col2rgb(hex1) + grDevices::col2rgb(hex2)) / 2),
     maxColorValue = 255
   )
+}
+
+scale_hex_brightness <- function(hex, scaling_factor) {
+  
+  if (all(scaling_factor == 1)) return(hex)
+
+  hsv_color <- hex %>%
+    grDevices::col2rgb() %>%
+    grDevices::rgb2hsv()
+
+  hsv_color["v", ] <- bound_num(hsv_color["v", ] * scaling_factor, 0, 1)
+  
+  grDevices::hsv(hsv_color["h", ], hsv_color["s", ], hsv_color["v", ])
 }